+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to
 - ✨(api) modify items/search endpoint to use indexed items in Find
 - 🐛(email) avoid trying to send emails if no provider is configured
 - ♻️(backend) improve mimetype detection
+- ♻️(backend) remove N+1 query patterns on items children view
 
 ### Fixed